Trinity packaging for I/O and Compute nodes.
Updated the Trinity packaging scripts to reflect the packages being
installed on the I/O and Compute nodes. SOS and Baler are not
being installed on these systems.

Also updated the list of packages to include those that are
to be installed on these systems and are not build-able on
non-Cray platforms.